requirements.txt
pandas
nltk
scikit-learn
streamlit

# DataLens
langchain
langchain-openai
langchain_experimental==0.0.49
python-dotenv
streamlit
openai
Tabulate
transformers
torch
ydata-profiling

#anthropic
langchain-anthropic

# gemini
google-generativeai